requirements.txt
# Core RL & ML
gymnasium==0.29.1
stable-baselines3==2.2.1
sb3-contrib>=2.2.1  # TD3, TQC, QRDQN
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
tensorboard>=2.15.0  # Para monitoramento

# Trading & Data
ccxt>=4.0.0
TA-Lib>=0.6.8

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0

# Visualization & Monitoring
matplotlib>=3.7.0
streamlit>=1.28.0

# Logging & Database
sqlalchemy>=2.0.0

# LLM & Sentiment Analysis
openai>=1.6.0  # GPT-4/GPT-3.5 para análise de sentimento
anthropic>=0.8.0  # Claude (alternativa ao OpenAI)
transformers>=4.35.0  # Modelos locais (FinBERT, BERT)
torch>=2.0.0  # Backend para transformers
newsapi-python>=0.2.7  # Coletar notícias globais
feedparser>=6.0.10  # RSS feeds (CoinDesk, etc)
beautifulsoup4>=4.12.0  # Web scraping
requests>=2.31.0  # HTTP requests
textblob>=0.17.1  # Análise de sentimento básica
vaderSentiment>=3.3.2  # Sentimento de mídia social
